Removed 2-second sleep before rcon disconnect
We used to sleep for 2 seconds before disconnecting from rcon to avoid a bug in
Minecraft 1.2.5/Bukkit-R4.0. If we disconnected from rcon (rather than the server
timing us out), we would trigger some sort of behavior where the server would
start allocating memory like crazy and lag out, eventually crashing.

This does not appear to be the case any longer in 1.2.5/Bukkit-R5.0, so I'm
removing the avoidance.